refactor: gather utility functions for
