+++ packages_rs/nextclade/src/io/mod.rs
@@ -1,7 +1,6 @@
 pub mod compression;
 pub mod concat;
 pub mod console;
-pub mod container;
 pub mod csv;
 pub mod dataset;
 pub mod errors_csv;
